+++ src/Sherlock/components/FacetInterface.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * User: Zachary Tong
+ * Date: 3/14/13
+ * Time: 6:16 AM
+ */
+
+namespace sherlock\components;
+
+/**
+ * Interface for facet components, always used in conjunction with the BaseComponent class
+ */
+interface FacetInterface {
+
+}